COMPLICATED PUSH
This is one of those push that is a bit complicated. Basically, I copied a lot of my old code from the last time I attempted this project.

Which means, stuff works, but barely.

Anyway, try not to fall back on this version.
However, that's probably what I'm gonna do. Since this version is actually useable.

Actually no, it's not even useable. I mean the storyboard is pretty much perfectly done, but the main program is still completely not working. I think I'll need to skip forward and see if I can finish setting up the API server first. Get that done and stuff should run much smoother.